run-as: Bump the size of the internal packages list buffer.
This is a back-port of https://review.source.android.com/#change,23480
to the gingerbread branch. Fix for b/4546601

This patch increases the size of the internal buffer used by run-as
to store the content of /data/system/packages.list from 8KB to 64KB.

It has been reported that, on some systems, 8KB was too small. This
resulted in a truncated file being loaded, and the inability to debug
native applications properly (either because the application was not
found in the list, or because the tool reported a 'corrupted
installation' due to BAD_FORMAT issues when parsing the truncated
file).

See http://code.google.com/p/android/issues/detail?id=16391

Change-Id: Ie0bfb5847f6fb1c2c0a836aee2347911f78049c1